% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_equal.R
\name{any_equal}
\alias{any_equal}
\title{Checks if any person has been sampled to buy themselves a present.}
\usage{
any_equal(df)
}
\arguments{
\item{df}{A data frame containing two columns}
}
\value{
TRUE if an individual is to buy for themselves and FALSE if everybody is buying for someone else
}
\description{
Checks if any person has been sampled to buy themselves a present.
}
\examples{
pairs_df <- data.frame(buyer = c("tom", "dick", "harry"), receiver = c("harry", "tom", "dick"))
any_equal(df = pairs_df)
}
