% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pairs.R
\name{check_pairs}
\alias{check_pairs}
\title{Check randomised pairs of present-buyers and receivers.}
\usage{
check_pairs(df)
}
\arguments{
\item{df}{A data frame containing two columns}
}
\value{
The same data frame if all pairs pass the checks or an error if any check fails
}
\description{
Check randomised pairs of present-buyers and receivers.
}
\examples{
pairs_df <- data.frame(buyer = c("tom", "dick", "harry"), receiver = c("harry", "tom", "dick"))
check_pairs(df = pairs_df)
}
