% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pairs.R
\name{create_pairs}
\alias{create_pairs}
\title{Create a data frame of buyers and receivers for a secret santa randomisation}
\usage{
create_pairs(people)
}
\arguments{
\item{people}{A vector of names}
}
\value{
A data frame containing a column with the original list of names and a randomised list
}
\description{
Create a data frame of buyers and receivers for a secret santa randomisation
}
\examples{
create_pairs(people = c("tom", "dick", "harry"))
}
