% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nickr_group.R
\name{nickr_group}
\alias{nickr_group}
\title{Check that a condition holds on groups of data between pipe stages.}
\usage{
nickr_group(.data, cond, msg = "nickr_row", active = TRUE,
  logger = stop)
}
\arguments{
\item{.data}{Incoming data (omitted if in pipe).}

\item{cond}{Condition written using column names and aggregation functions.}

\item{msg}{User message to display if test fails.}

\item{active}{Is this check turned on (default TRUE). Set FALSE to disable test (e.g., in production).}

\item{logger}{Function to call with message (e.g., `warning` or `stop`).}
}
\value{
Input data without modification.
}
\description{
Expressions can use special functions like \code{n()} that are allowed in
\code{summarize} calls. This function should only be put in pipes after
\code{group_by} has been called (so that groups exist).
}
\examples{
\dontrun{
data \%>\%
  # Preliminary processing.
  mutate(...) \%>\%

  # Aggregation.
  group_by(person_id) \%>\%

  # Warn if there are less than 5 records in any group in production.
  nickr_group(n() >= 5, msg = "Should have at least 5 records",
              active = IN.PRODUCTION, logger = warning)

  # Further processing on unmodified data.
  summarize(...)
}
}
