% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nickr_row.R
\name{nickr_row}
\alias{nickr_row}
\title{Check that a condition holds on the rows of data between pipe stages.}
\usage{
nickr_row(.data, cond, msg = "nickr_row", logger = stop,
  active = TRUE)
}
\arguments{
\item{.data}{Incoming data (omitted if in pipe).}

\item{cond}{Condition written as one-sided formula with `.x` as the row and `.y` as the row index, e.g., `~ .x$a < .x$b`}

\item{msg}{User message to display if test fails.}

\item{logger}{Function to call with message (e.g., `warning` or `stop`).}

\item{active}{Is this check turned on (default TRUE). Set FALSE to disable test (e.g., in production).}
}
\value{
Input data without modification.
}
\description{
Check that a condition holds on the rows of data between pipe stages.
}
