% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_pairs.R
\name{sample_pairs}
\alias{sample_pairs}
\title{Randomise secret santas to generate pairs of gift buyers and receivers.}
\usage{
sample_pairs(people)
}
\arguments{
\item{people}{A vector of names}
}
\value{
A data frame containing a column with the original list of names and a randomised list
}
\description{
Randomise secret santas to generate pairs of gift buyers and receivers.
}
\examples{
sample_pairs(people = c("tom", "dick", "harry"))
}
