% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nickr.R
\docType{package}
\name{nickr}
\alias{nickr}
\alias{nickr-package}
\title{Check conditions in pipelines, logging messages if tests fail.}
\description{
Check conditions in pipelines, logging messages if tests fail.
}
\author{
Greg Wilson, \email{greg.wilson@rstudio.com}
}
