% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nickr_col.R
\name{nickr_col}
\alias{nickr_col}
\title{Check that a condition holds on the columns of data between pipe stages.}
\usage{
nickr_col(.data, ..., msg = "nickr", logger = stop, active = TRUE)
}
\arguments{
\item{.data}{Incoming data (omitted if in pipe).}

\item{...}{Column-wise condition expressions to test.}

\item{msg}{User message to display if test fails.}

\item{logger}{Function to call with message (e.g., `warning` or `stop`).}

\item{active}{Is this check turned on (default TRUE). Set FALSE to disable test (e.g., in production).}
}
\value{
Input data without modification.
}
\description{
Check that a condition holds on the columns of data between pipe stages.
}
