% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nickr_row.R
\name{nickr_row}
\alias{nickr_row}
\title{Check that a condition holds on the rows of data between pipe stages.}
\usage{
nickr_row(.data, cond, msg = "nickr_row", active = TRUE,
  logger = stop)
}
\arguments{
\item{.data}{Incoming data (omitted if in pipe).}

\item{cond}{Condition written using column names, \code{.ri}, and \code{.row}.}

\item{msg}{User message to display if test fails.}

\item{active}{Is this check turned on (default TRUE). Set FALSE to disable test (e.g., in production).}

\item{logger}{Function to call with message (e.g., `warning` or `stop`).}
}
\value{
Input data without modification.
}
\description{
This is used much less frequently than \code{nickr_col}, since most
tests are both easier to express and more efficient to run on columns,
but it does have occasional uses.  See \code{nickr_col} for examples
of using the \code{active} and \code{logger} arguments.
}
\details{
Each row is augmented with \code{.ri} for the row index and \code{.row}
to hold the entire original row as a list.
}
\examples{
\dontrun{
data \%>\% mutate(...) \%>\%

  # Check that at most one field is zero in any row.
  nickr_row(sum(as.double(.row) == 0) <= 1,
            msg = "at most one zero per row") \%>\%

  # Further processing on unmodified data.
  filter(...)
}
}
